Refactor and group code
