+++ godtools/Views/TractElements/TractCallToAction.swift
@@ -62,26 +62,8 @@ class TractCallToAction: BaseTractElement {
     
     // MARK: - Setup
     
-    override func setupView(properties: Dictionary<String, Any>) {
-        super.setupView(properties: properties)
+    override func loadStyles() {
         addArrowButton()
-    }
-    
-    func addArrowButton() {
-        let xPosition = self.buttonXPosition
-        let yPosition = (self.height - self.buttonSizeConstant) / 2
-        let origin = CGPoint(x: xPosition, y: yPosition)
-        let size = CGSize(width: self.buttonSizeConstant, height: self.buttonSizeConstant)
-        let buttonFrame = CGRect(origin: origin, size: size)
-        let button = UIButton(frame: buttonFrame)
-        let image = UIImage(named: "right_arrow_blue")
-        button.setBackgroundImage(image, for: UIControlState.normal)
-        button.addTarget(self, action: #selector(moveToNextView), for: UIControlEvents.touchUpInside)
-        self.addSubview(button)
-    }
-    
-    func moveToNextView() {
-        NotificationCenter.default.post(name: .moveToNextPageNotification, object: nil)
     }
     
     override func textStyle() -> TractTextContentProperties {
@@ -98,6 +80,21 @@ class TractCallToAction: BaseTractElement {
                       y: self.yPosition,
                       width: self.width,
                       height: self.height)
+    }
+    
+    // MARK: - UI
+    
+    func addArrowButton() {
+        let xPosition = self.buttonXPosition
+        let yPosition = (self.height - self.buttonSizeConstant) / 2
+        let origin = CGPoint(x: xPosition, y: yPosition)
+        let size = CGSize(width: self.buttonSizeConstant, height: self.buttonSizeConstant)
+        let buttonFrame = CGRect(origin: origin, size: size)
+        let button = UIButton(frame: buttonFrame)
+        let image = UIImage(named: "right_arrow_blue")
+        button.setBackgroundImage(image, for: UIControlState.normal)
+        button.addTarget(self, action: #selector(moveToNextView), for: UIControlEvents.touchUpInside)
+        self.addSubview(button)
     }
 
 }
