+++ godtools/Views/TractElements/TractCallToActionActions.swift
@@ -1,0 +1,17 @@
+//
+//  TractCallToActionActions.swift
+//  godtools
+//
+//  Created by Devserker on 5/29/17.
+//  Copyright © 2017 Cru. All rights reserved.
+//
+
+import Foundation
+
+extension TractCallToAction {
+    
+    func moveToNextView() {
+        NotificationCenter.default.post(name: .moveToNextPageNotification, object: nil)
+    }
+    
+}
